Fix #[cfg] directives for preview feature
